==> analysis/scripts/2_results_labeling.ipynb <==
import pandas as pd
import numpy as np

**Labeling open text questions from the study results.**
- input: formatted results export from UXtweak study
- output: auto-labeled results, some still need to be labeled manually

# Result labeling

results = pd.read_csv("../data/results_unlabeled.csv", index_col=0)

results.info()

results.head()

## Automatic labeling

results.company = results.company.astype(str)
results['company_eval'] = 0

results.purpose = results.purpose.astype(str)
results.insert(6, 'purpose_eval', 0)

results.detail = results.detail.astype(str)
results['detail_eval'] = 0

**Company's name labeling.**

for index, row in results.iterrows():
    if(
        row.company.strip().lower() in [
            'whistles','whistle', 'whistless', 'whistels', 'wistles', 'whistlers', 'whilstles'] + [
            'i-care',  'i care', 'icare',  'i -care', 'i- care', 'i - care'] + [
            'hillsport', 'hill sport', 'hilsport', 'hilsport', 'hill-sport', 'hills sport', 'hillspot', 'hill sports', 'hill spot'] + [
            'weathery', 'weatherly', 'weather', 'wethery', 'weathry', 'weatherby', 'wearthy'] + [
            'mega zoo', 'megazoo', 'mega-zoo', 'meg zoo', 'mego zoo', 'maga zoo', 'mega zoom'] + [
            'tech live', 'tech-live', 'techlive', 'tech - live', 'techlife', 'tech-life', 'tech life', 'tech live (i think)', 'live tech'
        ]
    ):
        results.loc[index, 'company_eval'] = 1
    elif(
        row.company.strip().lower() in [
            'eye care', 'care', 'ice care', 'i-clear',
            'hillwear', 'sports', 'hillstart', 'hill sort', 'hill',
            'zoo', 'mini zoo', 'repti zoo', 'mega vets', 'moo zoo', 'zoofood',
            'weather office',
            'tech', 'tech - something', 'tech page', 'tech lite'
        ]
    ):
        results.loc[index, 'company_eval'] = 0.5
    else:
        results.loc[index, 'company_eval'] = 0

**Company's purpose labeling.**

for index, row in results.iterrows():
    if(
        row.task == 1 and any(word in row.purpose.strip().lower() for word in [
            'clothing', 'clothes', 'sustainable fashion', 'cloths'
        ]) or
        row.task == 2 and any(word in row.purpose.strip().lower() for word in [
            'glasses', 'eyewear', 'eye wear', 'eye products', 'eye care products', 'eye services', 'gkasses',
            'vision related services', 'eyeware',  'opticians', 'lenses'
        ]) or
        row.task == 3 and any(word in row.purpose.strip().lower() for word in [
            'clothes', 'clothing', 'sportswear', 'outdoor', 'mountain sports', 'activity equipment', 'sports attire'
            'sportwear', 'hiking', 'sports wear', 'sporting equipment', 'gear', 'activewear', 'sportwear', 'sports ware',
            'active wear', 'sports attire'
        ]) or
        row.task == 4 and any(word in row.purpose.strip().lower() for word in [
            'forecast', 'weather', 'wheather'
        ]) or
        row.task == 5 and any(word in row.purpose.strip().lower() for word in [
            'pet food', 'pet supplies', 'for animals', 'pet products', 'pet related products', 'animal food',
            'animal products', 'pet items', 'pet supples', 'related to pet-keeping', 'supplies for pets',
            'animal accessories', 'animal supplies', 'stuff for pets', 'pet related', 'pet accessories',
            'pet needs', 'pet-related', 'products for a variety of pets', 'pet care products', 'varoety of things for pets',
            'items for pets', 'they sell food, and toys for pets'
        ]) or
        row.task == 6 and any(word in row.purpose.strip().lower() for word in [
            'tech news', 'technology news', 'about tech', 'tech information', 'tech industry', 'gadgets and technology',
            'latest tecnological', 'on tech', 'of tech', 'tech related', 'modern technologies', 'technology related', 
            'tech blog', 'technology community', 'for tech', 'tech knowledge', 'relating to technology',
            'news about itech', 'tech info', 'technology sector', 'technology info', 'tech articles'
        ])
    ):
        results.loc[index, 'purpose_eval'] = 1
    elif(
        row.task == 2 and row.purpose.strip().lower() in [
            'eye things and holiday things', 'eyesight'
        ] or
        row.task == 3 and row.purpose.strip().lower() in [
            'extreme winter sport services', 'backpacks and running shoes', 'it\'s about trekking.'
        ] or
        row.task == 5 and any(word in row.purpose.strip().lower() for word in [
            'animal feed', 'animals', 'animal stuff', 'petfood', 'bio sand'
        ]) or
        row.task == 6 and any(word in row.purpose.strip().lower() for word in [
            'tech', 'news', 'articles', 'online magazine'
        ])
    ):
        results.loc[index, 'purpose_eval'] = 0.5
    else:
        results.loc[index, 'purpose_eval'] = 0

**Web page's detail labeling.**

for index, row in results.iterrows():
    if(
        row.task == 1 and row.detail.strip().lower() in [
            'hedgehog', 'hedghog', 'badger/hedgehog'
        ] or
        row.task == 2 and row.detail.strip().lower() in [
            '2', '2?', 'i think there were two children.', '2 ?', 'two', '2 kids two adults i think'
        ] or
        row.task == 3 and row.detail.strip().lower() in [
            '2020', '2020 summer/fall', '2020 I think'
        ] or
        row.task == 4 and row.detail.strip().lower() in [
            'thursday', 'i think maybe thursday? but i\'m not sure why i think that.', 'thursday?'
        ] or
        row.task == 5 and row.detail.strip().lower() in [
            'paw prints', 'dog treats', 'dog paw?', 'bones', 'paws', 'a paw', 'dog bones', 'dogbone?',
            'dog bone', 'paw prints?', 'dog paws', 'paw and circle', 'paw prints?', 'dog paw', 
            'i think i saw a paw print.', 'maybe a paw print?', 'paw prints, bones and leaves?', 
            'i think it was paw prints. it (as best as i can remember) was a white and gold box with black pawprints.',
            'bones (for dogs)', 'dog treats such as bones etc?', 'bones and the company logo', 'a paw print',
            'bone', 'animal footprints', 'cat paw/print'
        ] or
        row.task == 6 and row.detail.strip().lower() in [
            'landscape', 'i think there was some sort of power plant', 'wind turbine', 'wind turbines',
            'power plant', 'electricity plant', 'alternative power source', 'tanks with a windmill', 'windmill', 
            'i\'m not sure i think it was a view of something like a power plant or such but could be wrong. it appeared to be outdoors and industrial in nature but again i could be wrong.',
            'it looked like a battery farm but i am not sure if this is the correct term. it looks like it is in a remote location with a turbine in the background'
        ]
    ):
        results.loc[index, 'detail_eval'] = 1
    elif(
        row.task == 2 and row.detail.strip().lower() in [
           '1 or 2', '2-3?', 'unsure - 2 or 3', "1 or 2,  but i can't recall.",
        ] or
        row.task == 3 and row.detail.strip().lower() in [
           '202-',
        ] or
        row.task == 4 and row.detail.strip().lower() in [
           'not sure but tuesday or thursday', 'all days of the week', 'everyday'
        ] or
        row.task == 5 and row.detail.strip().lower() in [
           'i think it was the characters of the brand name mega zoo', 'circle?', 'circle',
           'the mega zoo logo i think', 'mega zoo'
        ] or
        row.task == 6 and row.detail.strip().lower() in [
           'battery farm', 'A power station', 'power station', 'an industrial plant - refinery or something similar'
        ]
    ):
        results.loc[index, 'detail_eval'] = 0.5
    else:
        results.loc[index, 'detail_eval'] = 0

results.columns

**Remove unecessary columns.**

results.company = results.company_eval
results.detail = results.detail_eval

results = results.drop(columns={
    'company_eval', 'detail_eval'
})

results.head()

## Preparation for manual labeling

results.insert(9, 'desc_eval_char', '-')
results.insert(9, 'desc_eval_elem', '-')
results.insert(9, 'desc_eval_comp', '-')

results.head()

**Export.**

results.to_csv("../data/results_to_label.csv")
