==> analysis/scripts/3_data_calculation.ipynb <==
import pandas as pd
import numpy as np
import re
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
import nltk
import json
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('omw-1.4')

**Calculating additional features to use in the analysis.**
- input: formatted results and respondents cognitive tasks data
- output: final results and respondents exports to be used for statistical analysis

# Calculating features

## Calculating new result features

**Load results.**

results = pd.read_csv("../data/results_labeled.csv", index_col=0)

results.info()

results.head()

**Add columns.**

results.insert(22, 'ATT', '')
results.insert(22, 'HQ', '')
results.insert(22, 'PQ', '')
results.insert(12, 'desc_eval_sum', '')
results.insert(9, 'desc_word_count_group', '')
results.insert(9, 'desc_word_count', '')
results.insert(9, 'desc_token_count_group', '')
results.insert(9, 'desc_token_count', '')
results.insert(9, 'desc_tokens', '')
results.insert(7, 'purpose_word_count', '')
results.insert(7, 'purpose_token_count', '')
results.insert(7, 'purpose_tokens', '')

results.info()

**Calculate number of words and number of full meaning words.**

lemmatizer = WordNetLemmatizer()

for index, row in results.iterrows():
    if(type(row.description) == str):
        tokens = word_tokenize(row.description.replace('/', ' ').replace('-', ' '))
        tokens_no_punct = list(filter(lambda x: x.isalpha(), [word.lower() for word in tokens if word.isalpha()]))
        tokens_no_stop = [word for word in tokens_no_punct if not word in stopwords.words()]
        tokens_final = list(map(lambda x: lemmatizer.lemmatize(x), tokens_no_stop))
        results.loc[index, 'desc_word_count'] = len(tokens_no_punct)
        results.loc[index, 'desc_token_count'] = len(tokens_final)
        results.loc[index, 'desc_tokens'] = ' '.join(tokens_final)
    else:
        results.loc[index, 'desc_word_count'] = 0
        results.loc[index, 'desc_token_count'] = 0
    if(type(row.purpose) == str):
        tokens = word_tokenize(row.purpose.replace('/', ' ').replace('-', ' '))
        tokens_no_punct = list(filter(lambda x: x.isalpha(), [word.lower() for word in tokens if word.isalpha()]))
        tokens_no_stop = [word for word in tokens_no_punct if not word in stopwords.words()]
        tokens_final = list(map(lambda x: lemmatizer.lemmatize(x), tokens_no_stop))
        results.loc[index, 'purpose_word_count'] = len(tokens_no_punct)
        results.loc[index, 'purpose_token_count'] = len(tokens_final)
        results.loc[index, 'purpose_tokens'] = ' '.join(tokens_final)
    else:
        results.loc[index, 'purpose_word_count'] = 0
        results.loc[index, 'purpose_token_count'] = 0

results.columns

results.loc[results.desc_word_count <= 11, 'desc_word_count_group'] = 'low' 
results.loc[(results.desc_word_count > 11) & (results.desc_word_count < 24), 'desc_word_count_group'] = 'med' 
results.loc[results.desc_word_count >= 24, 'desc_word_count_group'] = 'high' 
results.groupby('desc_word_count_group').size()

results.loc[results.desc_token_count <= 5, 'desc_token_count_group'] = 'low' 
results.loc[(results.desc_token_count > 5) & (results.desc_token_count < 13), 'desc_token_count_group'] = 'med' 
results.loc[results.desc_token_count >= 13, 'desc_token_count_group'] = 'high' 
results.groupby('desc_token_count_group').size()

**Remove unecessary columns.**

results.purpose = results.purpose_eval

results = results.drop(columns={
    'description', 'purpose_eval'
})

results.head()

**Means of Attrak-Diff results.**

results['PQ'] = (results.PQ1 + results.PQ2 + results.PQ3 + results.PQ4) / 4
results['HQ'] = (results.HQ1 + results.HQ2 + results.HQ3 + results.HQ4) / 4
results['ATT'] = (results.ATT1 + results.ATT2) / 2

**Sum of webpage elements and characteristics.**

results['desc_eval_sum'] = results['desc_eval_elem'] + results['desc_eval_char']

results.columns

**Replace faulty results.**

for i in range(1, 7):
    results.loc[i + 677, 'time'] = np.floor(results[results.task == i].time.mean())

**Insert design complexities.**

complexities = {
    1: 0.22,
    2: 2,
    3: 2.97,
    4: 5.27,
    5: 7.5,
    6: 8
}

results.insert(2, 'complexity', 0)
for index, row  in results.iterrows():
    results.loc[index, 'complexity'] = complexities[row.task]

**Remove respondent 40**

results = results[(results.respondent != 40) | (results.interval != '2s')]

**Export.**

results.to_csv("../data/results.csv")

## Respondent cognitive tests calculations

**Load data.**

respondents = pd.read_csv("../data/respondents_to_calculate.csv", index_col=0)

**Remove respondents that did not finish and respondent 40**

respondents = respondents[(respondents.status == 'completed')]

respondents = respondents[(respondents.respondent != 40) | (respondents.interval != '2s')]

respondents.shape

respondents.columns

respondents.insert(19, 'pst_press_speed', 0)
respondents.insert(18, 'pst_levels_correct_group', 0)
respondents.insert(17, 'pst_mistakes_group', '')
respondents.insert(17, 'pst_mistakes', 0)
respondents.insert(17, 'pst_levels_group', '')
respondents.insert(16, 'vmt_press_len_group', '')
respondents.insert(16, 'vmt_press_len', 0)
respondents.insert(16, 'vmt_click_speed_group', '')
respondents.insert(16, 'vmt_click_speed', 0)
respondents.insert(15, 'vmt_levels_group', '')

**Calculate press lengths and click speeds.**

for index, respondent in respondents.iterrows():
    if(type(respondent.vmt_clicks) == str):
        levels = json.loads(respondent.vmt_clicks)
        sum_time = 0
        sum_press_time = 0
        clicks = 0
        between_clicks = 0
        for level in levels:
            last_time = 0
            for click in level:
                if(click['type']=='mouseDown'):
                    sum_time += (click['time'] - last_time) if last_time != 0 else 0
                    clicks += 1
                    between_clicks += 1 if last_time != 0 else 0
                    last_time = click['time']
                if(click['type']=='mouseUp'):
                    sum_press_time += (click['time'] - last_time)
        
        respondents.loc[index, 'vmt_click_speed'] = (between_clicks / sum_time ) * 1000
        respondents.loc[index, 'vmt_press_len'] = sum_press_time / clicks

for index, respondent in respondents.iterrows():
    if(type(respondent.pst_clicks) == str):
        presses = json.loads(respondent.pst_clicks)
        sum_time = presses[len(presses) - 1] - presses[0]
        respondents.loc[index, 'pst_press_speed'] = (len(presses) / sum_time ) * 1000

**Assign groups according to cognitive tasks**

respondents['pst_mistakes'] = respondents['pst_levels'] - respondents['pst_levels_correct']
respondents['pst_mistakes'] = respondents['pst_mistakes']

respondents.loc[respondents.vmt_levels <= 5, 'vmt_levels_group'] = 'low' 
respondents.loc[(respondents.vmt_levels > 5) & (respondents.vmt_levels < 8), 'vmt_levels_group'] = 'med' 
respondents.loc[respondents.vmt_levels >= 8, 'vmt_levels_group'] = 'high' 
respondents.groupby('vmt_levels_group').size()

respondents.loc[respondents.pst_levels <= 30, 'pst_levels_group'] = 'low' 
respondents.loc[(respondents.pst_levels > 30) & (respondents.pst_levels < 38), 'pst_levels_group'] = 'med' 
respondents.loc[respondents.pst_levels >= 38, 'pst_levels_group'] = 'high' 
respondents.groupby('pst_levels_group').size()

respondents.loc[respondents.pst_levels_correct <= 27, 'pst_levels_correct_group'] = 'low' 
respondents.loc[(respondents.pst_levels_correct > 27) & (respondents.pst_levels_correct < 36), 'pst_levels_correct_group'] = 'med' 
respondents.loc[respondents.pst_levels_correct >= 36, 'pst_levels_correct_group'] = 'high' 
respondents.groupby('pst_levels_correct_group').size()

respondents.loc[respondents.pst_mistakes <= 0, 'pst_mistakes_group'] = 'low' 
respondents.loc[(respondents.pst_mistakes > 0) & (respondents.pst_mistakes < 3), 'pst_mistakes_group'] = 'med' 
respondents.loc[respondents.pst_mistakes >= 3, 'pst_mistakes_group'] = 'high' 
respondents.groupby('pst_mistakes_group').size()

respondents.loc[respondents.vmt_click_speed <= 1.34, 'vmt_click_speed_group'] = 'low' 
respondents.loc[(respondents.vmt_click_speed > 1.34) & (respondents.vmt_click_speed < 1.7), 'vmt_click_speed_group'] = 'med' 
respondents.loc[respondents.vmt_click_speed >= 1.7, 'vmt_click_speed_group'] = 'high' 
respondents.groupby('vmt_click_speed_group').size()

respondents.loc[respondents.vmt_press_len <= 82.3, 'vmt_press_len_group'] = 'low' 
respondents.loc[(respondents.vmt_press_len > 82.3) & (respondents.vmt_press_len < 112), 'vmt_press_len_group'] = 'med' 
respondents.loc[respondents.vmt_press_len >= 112, 'vmt_press_len_group'] = 'high' 
respondents.groupby('vmt_press_len_group').size()

**Remove uneccessary columns.**

respondents = respondents.drop(columns={
    'vmt_clicks', 'pst_clicks'
})

respondents.columns

respondents.head()

**Export.**

respondents.to_csv("../data/respondents.csv")
